Fix Mod Menu bug
The bug is related to the `ModMenuApi#getProvidedConfigScreenFactories` interface, which works every other time. The fix replaces the use of this interface with direct access to a variable that stores all configuration screens. With this fix, Elegantia will work regardless of whether the Mod Menu is in the list of mods